pyproject.toml
[project]
name = "invoice-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.2.1",
    "pydantic>=2.11.5",
    "pydantic-ai>=0.2.12",
    "pydantic-graph>=0.2.12",
    "pydantic-settings>=2.9.1",
    "pypdfium2>=4.30.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
